Clean up error naming (#11)
