+++ src/converters/RpcConverters.ts
@@ -11,7 +11,7 @@ import {
     RpcTypedData,
 } from '@azure/functions-core';
 import { isLong } from 'long';
-import { InternalException } from '../utils/InternalException';
+import { AzFuncSystemError } from '../errors';
 
 /**
  * Converts 'ITypedData' input from the RPC layer to JavaScript types.
@@ -100,7 +100,7 @@ export function toNullableBool(nullable: boolean | undefined, propertyName: stri
     }
 
     if (nullable != null) {
-        throw new InternalException(
+        throw new AzFuncSystemError(
             `A 'boolean' type was expected instead of a '${typeof nullable}' type. Cannot parse value of '${propertyName}'.`
         );
     }
@@ -132,7 +132,7 @@ export function toNullableDouble(
     }
 
     if (nullable != null) {
-        throw new InternalException(
+        throw new AzFuncSystemError(
             `A 'number' type was expected instead of a '${typeof nullable}' type. Cannot parse value of '${propertyName}'.`
         );
     }
@@ -152,7 +152,7 @@ export function toRpcString(nullable: string | undefined, propertyName: string):
     }
 
     if (nullable != null) {
-        throw new InternalException(
+        throw new AzFuncSystemError(
             `A 'string' type was expected instead of a '${typeof nullable}' type. Cannot parse value of '${propertyName}'.`
         );
     }
@@ -174,7 +174,7 @@ export function toNullableString(nullable: string | undefined, propertyName: str
     }
 
     if (nullable != null) {
-        throw new InternalException(
+        throw new AzFuncSystemError(
             `A 'string' type was expected instead of a '${typeof nullable}' type. Cannot parse value of '${propertyName}'.`
         );
     }
@@ -204,7 +204,7 @@ export function toNullableTimestamp(
                 };
             }
         } catch {
-            throw new InternalException(
+            throw new AzFuncSystemError(
                 `A 'number' or 'Date' input was expected instead of a '${typeof dateTime}'. Cannot parse value of '${propertyName}'.`
             );
         }
