--- src/utils/InternalException.ts
@@ -1,6 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License.
-
-export class InternalException extends Error {
-    isAzureFunctionsInternalException = true;
-}
